+++ framework/py/flwr/server/app.py
@@ -73,8 +73,8 @@ from flwr.server.serverapp.app import flwr_serverapp
 from flwr.simulation.app import flwr_simulation
 from flwr.supercore.ffs import FfsFactory
 from flwr.supercore.object_store import ObjectStoreFactory
-from flwr.superexec.app import load_executor
-from flwr.superexec.exec_grpc import run_exec_api_grpc
+from flwr.superlink.executor.app import load_executor
+from flwr.superlink.servicer.exec.exec_grpc import run_exec_api_grpc
 
 from .superlink.fleet.grpc_adapter.grpc_adapter_servicer import GrpcAdapterServicer
 from .superlink.fleet.grpc_rere.fleet_servicer import FleetServicer
