+++ framework/py/flwr/superlink/servicer/exec/exec_license_interceptor_test.py
@@ -36,7 +36,7 @@ from flwr.proto.exec_pb2 import (  # pylint: disable=E0611
     StreamLogsRequest,
 )
 from flwr.supercore.license_plugin import LicensePlugin
-from flwr.superexec.exec_license_interceptor import ExecLicenseInterceptor
+from flwr.superlink.servicer.exec.exec_license_interceptor import ExecLicenseInterceptor
 
 HandlerContinuation = Callable[[grpc.HandlerCallDetails], grpc.RpcMethodHandler]
